Raise ExternalRequestError instead of HTTPError
Raise ExternalRequestError from HTTPService instead of HTTPError.

This fixes the reporting of failed HTTP requests by HTTPService to
Sentry (adding details of the error response to Sentry) and removes an
unnecessary class (we don't need both HTTPError and
ExternalRequestError).

When HTTPService was written (quite recently) a new HTTPError class was
added for it to raise:

    class HTTPError(ServiceError):
        """A problem with an HTTP request to an external service."""

        def __init__(self, response=None):
            super().__init__(response)
            self.response = response

I think this was probably a mistake. There was already an existing
ExternalRequestError class in the code and that class contains code for
stringifying the exception nicely for Sentry:

    class ExternalRequestError(ServiceError):
        def __init__(self, explanation=None, response=None, details=None):
            super().__init__()
            self.explanation = explanation
            self.response = response
            self.details = details

        def __str__(self):
            if self.response is None:
                return self.explanation

            # Log the details of the response. This goes to both Sentry and the
            # application's logs. It's helpful for debugging to know how the
            # external service responded.
            parts = [
                self.explanation + ":",
                str(self.response.status_code or ""),
                self.response.reason,
                self.response.text,
            ]
            return " ".join([part for part in parts if part])

HTTPService should have used ExternalRequestError all along.

The name HTTPError is kind of nicer than ExternalRequestError (it's
shorter for one thing) but I think it's probably best to stick with
ExternalRequestError to avoid confusion with
pyramid.httpexceptions.HTTPError or requests.HTTPError.