Reduce the size of the `coverage` tox environment
We don't need to install all of the test requirements in the `coverage`
environment, only `coverage` needs to be installed. This makes creating
the coverage environment much quicker, updating it both quicker and less
frequent, and means the environment takes less space on disk.

CI, `make coverage` and `make sure` will be faster.