Add `resourcekey` parameter to Google Drive PDF URLs
If a user selects a PDF file for an assignment which has a "resource
key" associated with it, capture this key from the Google
Picker-provided file metadata and add it to the file download URL as a
`resourcekey` query parameter. This mirrors the updated download URL for
affected files if you use the "Download" action for the file in the Google Drive UI.

The `resourceKey` field on file objects returned by the picker API is
not documented at https://developers.google.com/picker/docs/reference#document but it is
present in my testing.

It would still be preferable for us not to generate this URL ourselves,
but this commit is an easy incremental change that captures the
necessary information for files affected by the security update.