Fix gap at top of modal layout
The manifestation of this issue is that there
was a .5rem "white stripe" at the top of the viewport
when a the modal here was rendered, and the modal
itself had a .5rem top margin (making it off-center
vertically).

The cause of this is that there is a vertical-rhytym
applied to the parent element. Modals render in the
DOM "where they are". Though both the modal and its
overlay are fixed-positioned (removed from layout
flow), they are present in the DOM wherever they
are rendered, and are thus vulnerable to inheriting
CSS styles.

A broader fix may involve re-evaluating whether modals
should "render into" a different element, but for
now wrapping the modal here in another <div> fixes the
immediate problem: the vertical-spacing is applied
only to immediate children of `.FilePickerApp__right-col`
and so this <div> (empty/invisible) "absorbs" that.
Thus the modal and its overlay no longer have a top margin.