Fix missing test coverage on services/oauth2_token.
This was broken inadvertently on b08857160cfedee385bb4937c79e15cc263e3ab0,
the sqla fix had the effect of, as stated in the removed comment, always
adding the object to the session and missing the creating part of save().

`oauth_token_in_db_or_not` is only used in this test and it doesn't need
the not in the DB OAuth2Token object so the fixture is changed to return
None instead of a not yet added to the session object.