Add `react-intl` and update build to accommodate
`react-intl` is a `react`-ecosystem package and as such imports
`react` (peer dependency). Update `browserify` transforms to alias
`react` imports to `preact/compat` during bundling. This needs to be
a global transform, and as such must use browserify's API: it is not
possible to do this in package.json configuration.

See https://preactjs.com/guide/v8/switching-to-preact/#aliasing-via-browserify

Note that the `aliasify` transform must come after `babelify` (which
is configured in package.json and is applied during the main browserify
bundling in `create-bundle.js`).