Skip tests that break when importing `react-intl` components anywhere
These tests start failing as soon as any component in the app imports
one of the `react-intl` components. The breakage seems to be specifically
in the `input.simulate('change')` in the `updateURL` helper function.
The handler is never invoked.

At first I thought this might be something to do with `useServices`
getting disturbed by the addition of an additional wrapping
`IntlProvider` but that does not seem to be the case.

Obviously this will need to be tracked down.