Add redirect page for MS OneDrive authentication (OAuth)
By default the OneDrive picker uses the page it was launched from as the redirect
URL for authentication. In the LMS context, the OneDrive picker opens a
window with the URL '/content_item_selection' and some AOuth specific parameters (`?oauth=....`). That causes the LMS to return a 404 `Page not found`.

OneDrive has a `redirectUri` option to launch the picker in a different
URL (see
https://docs.microsoft.com/en-us/onedrive/developer/controls/file-pickers/js-v72/open-file?view=odsp-graph-online).
This URL must be registered on the app's registration portal as a
redirect URL.

The redirect page for authentication must content the OneDrive JS
bundle.