Don't wrap HAPIError in HTTPInternalServerError
I'm not sure why the code does this: `LTIHService` doesn't need to raise
an `HAPIError` wrapped in an `HTTPInternalServerError`. It can just
raise the `HAPIError` directly and the exception view can catch
`HAPIError`.

This is easier to follow because you can just see that the `HAPI`
service raises `HAPIError` and the exception view catches `HAPIError`
and you can see what the exception view does with that error. Rather
than having to know that the `HAPI` service raises `HAPIError` which
`LTIHService` catches and turns into an `HTTPInternalServerError` which
is then caught by the exception view for `HTTPServerError` (a parent
class of `HTTPInternalServerError`).

Note that the exception view previously caught `HTTPServerError` and all
its subclasses, which is all the Pyramid 5xx `httpexception`'s. But I
think the `HAPIError` wrapped in an `HTTPInternalServerError` was the
only case where any `HTTPServerError` was actually being raised. So I
don't think there's any unintended change in behaviour here.