Don't register non-API exception views for API requests
Fixes https://github.com/hypothesis/lms/issues/3148

Use a `path_info=not_("^/api/.*")` view predicate to ensure that the
non-API exception views are never used for API requests, and a similar
`path_info="^/api/.*"` predicate to ensure that API exception views are
_only_ used for API requests.

The exception views in `views/exceptions.py` are only intended to be
used for non-API requests but in fact most of them are registered for
all requests not just non-API requests. If one of these exceptions was
raised during an API request, and not caught by a more
specifically-configured exception view in `views/api/exceptions.py`,
then it would be caught by a `views/exceptions.py` exception view. This
would result in an HTML error page being sent as in API response.

Fix this by registering all the exception views in `views/exceptions.py`
for `/api/*` requests only.

To do this I indented the exception views into a class so that the
`path_info` view predicate could go in a `@view_defaults` decorator.
This avoids repeating the `path_info` predicate and also means that
it'll get applied to any new views that're added to this class in
future.

I also moved the `DEFAULT_RENDERER` constant into the `@view_defaults`.

The view methods themselves needed some light rewriting now that they're
part of a class, they need to use `self`. Other than that this commit
shouldn't change the views themselves at all.

The tests also needed to be adjusted as the views need to be called
differently now that they're methods of a class, and I rewrote the tests
to use pytest fixtures in the typical way rather than using a subclass.
This is less novel and makes the tests easier to understand and to
change.