Rename "reused tool GUID" to "reused consumer key"
The thing being reused here is called a "consumer key" (or "OAuth
consumer key") in LTI not a "tool GUID". The LTI spec actually contains
something else called a "tool GUID" (or "tool consumer instance GUID")
so it's probably best not to use "tool GUID" on place of "consumer key".