Make `ErrorDisplay` render `children`
Make `ErrorDisplay` able to render children, and thus be able to render
everything pertinent to an error. This will allow for more straightforward
overflow and scroll control as we transition to shared components.

We have to manage the display of error information in a wild assortment
of contexts and viewports, so it's easier to manage if error UI is more
encapsulated.