Take resource_link_id from query params when speed grading
The `resource_link_id` provided by canvas on launches done while using
speedgrading is incorrectly set to the value of "context_id` (the course
ID). This is not causing any problems (apart from some orphan
assignments
on the DB) because we are using the `url` and `file_id` URL params to
get the assignment details.

This would become a problem once we switch to reading the assignments
info from the DB instead of the URL like we do for other LMS so it needs
to be fixed.