Fix crashes in ExternalRequestError.__str__()
`ExternalRequestError.__str__()` is based on `self.explanation` and
`self.response`. Both of these are optional arguments to
`ExternalRequestError.__init__()` so there are four possible cases:

1. Both `response` and `explanation` are `None`:
   `ExternalRequestError()`
2. `explanation` is a string but `response` is `None`:
   `ExternalRequestError("Something went wrong")`
3. `responses` is a `Response` but `explanation` is `None`:
   `ExternalRequestError(response=response)`
4. `explanation` is a string and `response` is a `Response`:
   `ExternalRequestError(explanation, response)`

Two of these cases were missing from the tests and were actually
resulting in `__str__()` crashing. Add tests and fix the cases.

I'd like to also change the format of `ExternalRequestError`'s string
representations (it's a bit of a mess) but that's for another PR.