Remove HTTPError, use ExternalRequestError instead
`HTTPError` is the exception class raised by `HTTPService` whenever
anything goes wrong with a server-to-server HTTP request.

`HTTPService` is almost always used when sending HTTP requests: `HAPI`,
`LTIOutcomesAPIClient`, `VitalSourceService` and `BlackboardAPIClient`
all use `HTTPService`. `CanvasAPIClient` is the one thing that sends
external HTTP requests without using `HTTPService` and that's only
because `CanvasAPIClient`'s code is unusual and can't be easily
refactored to work like the others do.

`HTTPError` is kind of the same thing as `ExternalRequestError` but with
fewer attributes and less functionality.

This commit deletes `HTTPError` and changes `HTTPService` to raise
`ExternalRequestError` instead.

As a result all the services that use `HTTPService` will now get better
error reporting: details of error responses from third-party services
will be sent to Papertrail and Sentry.