Make sure ExternalRequestError's always have an explanation
Make sure that `ExternalRequestError` objects always have an
`explanation` attribute, defaulting to `"External request failed"`.

The exception view uses the `explanation` attribute as the `message` in
the API's JSON error response:

https://github.com/hypothesis/lms/blob/1b293991e21659d963fb65c9a05535e6303b8316/lms/views/api/exceptions.py#L106

API error responses have to have either an `error_code` or a `message`,
otherwise the frontend will treat them as an access token error and show
the authorization dialog instead of the error dialog:

https://github.com/hypothesis/lms/blob/1b293991e21659d963fb65c9a05535e6303b8316/lms/views/api/exceptions.py#L46-L48

Currently if a view raises an `ExternalRequestError` with no `message`
(`ExternalRequestError()` or `ExternalRequestError(response=response)`)
the `external_request_error()` exception view will end up sending an
error response with a JSON body of `{}` which results in an
authorization dialog not an error dialog.

Fix this by making sure that `ExternalRequestError`'s always have a
`message`.