Load JS bundles as ES modules
The Hypothesis client now relies on loading scripts as native ES modules, so it
makes sense to do the same in the LMS app.

There are no immediate benefits from this, but it will make loading
functionality on-demand (using `import()`) or code splitting easier in future.

As part of this, the baseline for the browser-compatibility check script
has been bumped to browsers which support `<script type="module">`.
Older browsers won't see any errors at all. Recent usage data indicates
that this will affect very few users.