Make pattern library use same CSS context as frontend apps
Update the jinja template and CSS sourcing for the local pattern
library such that the resulting CSS stack is the same as what's used
in the front end apps. This means that WYSIWY(actually)G for patterns
in the pattern library, and will be necessary to be able to add some
local-app pages to the pattern library (e.g. with error state examples).