Add some response details to error dialogs
When there's a problem with a server-to-server request to a third-party
service (an `ExternalRequestError`) add some details of the error
response from the third-party to the error dialog that we show to users.

This is to assist in debugging (for example when users send us
screenshots of the error dialog).

This commit adds the HTTP status code and reason of the response to the
error dialog. We also discussed adding the requested URL (minus query
params for security reasons) but this commit doesn't do that yet.

We don't want to show the response body to users for security reasons.

This partially addresses https://github.com/hypothesis/lms/issues/2998
but doesn't entirely fix it yet.