Give preference to `error.errorMessage` over `error.message` in UI
When showing details about an error to users in `ErrorDisplay`, use
`error.errorMessage` if it is present (even if empty), preferring it
over `error.message`. This prevents displaying an `error.message` that
is not useful to the user, and allows the backend to designate what
gets shown as a message in these cases.

Simplify `ErrorDisplay` logic accordingly.

Fixes #3259