Provide link that pre-populates support form with error information
Update the way we display errors to users such that there is a link to
create a new support ticket, and the form at that URL is pre-filled
with relevant details about the current error.

There are several possible combinations of information we might have
about a given error state, so try to populate this form usefully given
whatever we have access to.

Fixes https://github.com/hypothesis/support/issues/240