Simplify the blackboard_groups_enabled property
If `LTILaunchResource.blackboard_groups_enabled` checks
`application_instance.product` (and returns `False` if
`tool_consumer_info_product_family_code` isn't `"BlackboardLearn"`) we
run the risk of the Blackboard Groups feature being disabled when it
should be enabled if there are Blackboard instance's that use a
different `tool_consumer_info_product_family_code`. The code is also
more complicated.

On the other hand if `LTILaunchResource.blackboard_groups_enabled` just
checks the value of `application_instance.settings.get("blackboard",
"groups_enabled")` with checking the product then the risk is that we'll
enable Blackboard Groups in a non-Blackboard LMS. This would result in
broken behaviour, but:

* It would only happen if a Hypothesis admin incorrectly enabled
  Blackboard Groups for a non-Blackboard application instance

* It could be fixed by support by using the admin pages to disable
  Blackboard Groups for the application instance

The resulting code is also simpler.

One of the benefits of simpler code is that it's easier to see all the
possible cases, hence I was able to spot that the unit tests were
missing the case where `application_instance.settings.get("blackboard",
"groups_enabled")` returns `None` and I added a test case for that.