Re-enable_lti_launch_mode on basic_lti_launch.configure_assignment
After the assignment is configure parts of the JSConfig might have
changed so we need to re-enable lti launch mode.

The existing case this fix is while configuring a groups assignments.
The first call to enable_lti_launch_mode on __init__ won't populate
the sync API but once the assignment is configured a second call will.

As parts of js_config are memoized we need to manually clear its cache.