Debounce fetching students grades
In the LMS grader, every time that the next or previous arrow is pressed it
makes a network request. This could cause a big number of meaningless
requests. In order to avoid this, we have debounced the requests of
student grades, so that only the last request made within the last 200
milliseconds is executed, the rest are ignored.

This causes a delay of 200 milliseconds also when a specific user is
selected via the dropdown.