Rename colors for states to be less ambiguous
Naming a color `success` or `error` means that some of the tailwind-
generated utility classnames are less than crystal clear in some
contexts. For example, `.text-error` doesn't clearly communicate that
it is setting foreground/text color.

For this reason, I've elected to change these to `green-success`,
`yellow-notice` and `red-error`. I believe this communicates more
clearly the meaning of these utility classes.

Once this is landed, we can make this change to the global preset
in `frontend-shared` and remove this local `tailwind.config` change.