Use `AuthorizationModal` and `ErrorModal` to handle group-related errors
Previously, any failures in proxied requests to the Blackboard or
Canvas groups APIs for a list of group sets would be treated as an
authorization error. The previous interface would prompt for re-auth
within the main interface; it could appear that one was going into an
infinite auth loop if the underlying problem was non-auth-related.

These changes handle both authorization and other types of errors, and
put error UI into a modal.