Raise if a course grouping has a parent
`generate_authority_provided_id()` raises if you try to generate an
`authority_provided_id` for a non-course grouping with no parent. For
good measure it should also raise if you try to generate one for a
_course_ grouping _with_ a parent: if `type_ == Grouping.Type.COURSE`
then the `parent` argument would not be used to it makes no sense to
call `generate_authority_provided_id()` in this way.