Tidy upsert_grouping_memberships() tests
Collapse the two separate tests into a single test that tests both that
it updates an existing grouping and that it inserts a non-existing
grouping.

Avoid the `with_course_memberships` fixture, makes the test easier to
understand.

Use two different types of grouping.

Actually assert about the contents of the memberships table, instead of
counting things.