Break the loop between `web`, `devdata` and `services`.
Docker compose appears to check for the existence of the env file
even if you aren't starting the web service and so will refuse to start
the services if it's missing. Even though it should not play a role.

This causes a day 1 problem where you can't bootstrap the service
if you don't already have devdata, and you can't get devdata without
having the services.

By putting the command outside we break the loop.