Improve OneDrive permission error
Currently, if MS SharePoint is setup in such a way that doesn't allow to
create public, links the OneDrive error message caused is not very
helpful.

We have created a new error type, `PickerPermissionError`, that is
thrown when such permission restrictions are encountered. The new error
message contains more information about the error and a link to
additional resources.