Rename some test classes
Rename and slightly rearrange some test classes:

* Rename `TestCanvasAPIClient` (which contained the tests for all the
  methods _except_ `get_token()`) to `TestCanvasAPIClientIntegrated` and
  add a docstring explaining that this test class differs from the rest
  of the tests in that it tests `CanvasAPIClient` in integration with
  its dependencies rather than in isolation

* This now enables me to rename `TestCanvasAPIClientGetToken` to just
  `TestCanvasAPIClient`. This was the motivation for this commit:
  because a future commit is going to add tests for other methods to
  this class (not just tests for `get_token()`)

* Fix the `TestCanvasAPIClient.authenticated_client` fixture to use
  `create_autospec()` rather than using patch-at-source. Patch-at-source
  can create problems, for an explanation see
  <https://github.com/hypothesis/lms/pull/2729> and
  <https://github.com/hypothesis/via/pull/222#discussion_r469865055>

* Inlined the `TestMetaBehavior` class into `TestCanvasAPIClientIntegrated` since this class also contains integrated tests

Explanation:

The tests for `CanvasAPIClient` differ from most of our tests: they're
not isolated unit tests in the way we normally do it.

The tests use a `canvas_api_client` fixture to construct the
`CanvasAPIClient` instance to test:

    @pytest.fixture
    def canvas_api_client(authenticated_client, file_service):
        return CanvasAPIClient(authenticated_client, file_service)

    https://github.com/hypothesis/lms/blob/079f75979eff6d5c5500432d5bb5d6a254296f9d/tests/unit/lms/services/canvas_api/client_test.py#L528-L530

As you can see the `canvas_api_client` fixture is based on two other
fixtures `authenticated_client` and `file_service`. `file_service` is a
mock instance of `FileService`, as usual. But `authenticated_client` is
a real instance of `AuthenticatedClient`, the fixture is defined in
`canvas_api`'s own `conftest.py` file:

    @pytest.fixture
    def authenticated_client(basic_client, oauth2_token_service):
        return AuthenticatedClient(
            basic_client=basic_client,
            oauth2_token_service=oauth2_token_service,
            client_id=sentinel.client_id,
            client_secret=sentinel.client_secret,
            redirect_uri=sentinel.redirect_uri,
        )

    https://github.com/hypothesis/lms/blob/079f75979eff6d5c5500432d5bb5d6a254296f9d/tests/unit/lms/services/canvas_api/conftest.py#L20-L28

The `oauth2_token_service` fixture that `authenticated_client` depends
on is again a normal mock fixture. But the `basic_client` fixture
(defined further up in the same `conftest.py` file) again returns a real
instance of `BasicClient`.

So the tests for `CanvasAPIClient` are actually testing
`CanvasAPIClient`, `AuthenticatedClient` and `BasicClient` together in
integration, rather than testing each class as a separate unit in
isolation as we would normally do.

But there is _one_ test class at the top of `client_test.py` that is
different from the rest of the tests in this file though:
`TestCanvasAPIClientGetToken` overrides the `authenticated_client`
fixture with one that returns a mock, so this class actually does test
`CanvasAPIClient` with a mock `authenticated_client` (the
`canvas_api_client` fixture will still be called but it'll be passed a
mock `authenticated_client`):

    class TestCanvasAPIClientGetToken:
        # This is the only test where we fake out the underlying class, because
        # this _one_ call is just a pass through.

        ...

        @pytest.fixture
        def authenticated_client(self, patch):
            return patch("lms.services.canvas_api._authenticated.AuthenticatedClient")

    https://github.com/hypothesis/lms/blob/079f75979eff6d5c5500432d5bb5d6a254296f9d/tests/unit/lms/services/canvas_api/client_test.py#L11-L25

This commit renames `TestCanvasAPIClientGetToken` to just
`TestCanvasAPIClient`: a standard test class that uses our normal
mock-based approach, and renames the _other_ test class (the one that
contains integrated tests) to `TestCanvasAPIClientIntegrated`. This will
make room for more mock-style tests to be added to `TestCanvasAPIClient`
in future.