Refresh expired access tokens from the frontend
Implement frontend-coordinated refresh [1] of expired access tokens for external
APIs (Canvas, Blackboard). When a frontend API call fails because due to expiry
of an access token for an external API, initiate a refresh from the frontend.

Initiating the request from the frontend has the advantage that a client can
prevent concurrent refreshes, at least from the same browser session within
the scope of a single page load.

[1] https://github.com/hypothesis/lms/issues/3667#issuecomment-1049063194