Move application_instance property to lti resource
Don't check if application_instance exists in LTIResource

When the properties of LTIResource are used we already have check the
presence of an application_instance.

If by some reason that wasn't true we'd want to fail loudly and not try
to hide it by disabling some features.

Use the tool_consumer_instance_guid value directly from AI