Use tool_consumer_instance_guid from DB
tool_consumer_instance_guid is an optional parameter in the LTI spec but
required in our app.

In general all LMS do send it as a request parameters so
ApplicationInstance.update_lms_data will set the right value on the DB.

For LMS that don't send it we'll allow the request params to be absent
but we'll rely on the application_instance row to have the
tool_consumer_instance_guid value pre-filled.