Support direct/unauthenticated launch of VitalSource Bookshelf reader instead of LTI
Support directly loading the VitalSource Bookshelf reader in an iframe instead
of using an LTI launch which involves a form submission. When using this launch
mode, the LTI user authentication will not be passed to VitalSource and they
will need to login using VitalSource credentials.

This is a stop-gap solution to VitalSource assignment authentication as part of
https://github.com/hypothesis/private-issues/issues/102.

We'll take the decision of which method to use (the old LTI  based one or this new one)
based on the `vitalsource_anon_launch` feature flag.