Use a complete fixture and then modify it, instead of a very active factory
Previously we had a complex set of fixtures which constructed the right fixture
for each test. Quite a few of the arguments were specific to individual tests
and put distance between the setup and the assertions of the tests.

Now there is a complete example generated, which individual tests then modify as
required inside the tests, which makes them more self contained.