Enable the new frontend refresh feature
Slack thread: https://hypothes-is.slack.com/archives/C1MA4E9B9/p1652200138759249

When we [removed the `frontend_refresh` feature flag](https://github.com/hypothesis/lms/pull/3806)
we removed it and _disabled_ the feature rather than enabling it: the
new frontend-based OAuth 2 refresh code is currently *not* being used on
master. The old backend-based refresh is in use instead. To verify this:

Canvas
------

1. Launch [localhost (make devdata) Canvas Files Assignment](https://hypothesis.instructure.com/courses/125/assignments/875)
   so that you get a Canvas access token in your DB

2. Invalidate the access token in your DB:
   `tox -qe dockercompose -- exec postgres psql -U postgres -c "UPDATE oauth2_token SET access_token = 'foo';"`

3. Open the dev tools _Network_ tab and launch
   [localhost (make devdata) Canvas Files Assignment](https://hypothesis.instructure.com/courses/125/assignments/875)
   again.

   The assignment should launch successfully _without_ showing you an
   authorization dialog and if you look in the dev tools you will _not_
   see a request to the new `/api/canvas/oauth/refresh` API: the access
   token was refreshed by the backend without frontend involvement.

If you repeat the same test on _this_ branch the same thing will happen
but you will see a request to the refresh API in dev tools.

Blackboard
----------

1. Go to [Developer Test Course with Original Course View](https://aunltd-test.blackboard.com/ultra/courses/_19_1/cl/outline)
   and launch **localhost (make devdata) Blackboard Files Assignment**
   so that you get an access token in your DB.

2. Invalidate the access token in your DB:
   `tox -qe dockercompose -- exec postgres psql -U postgres -c "UPDATE oauth2_token SET access_token = 'foo';"`

3. Open the dev tools _Network_ tab and launch the same assignment again.

   As in Canvas the assignment should launch successfully _without_
   showing you an authorization dialog and if you look in the dev tools
   you will _not_ see a request to the new
   `/api/blackboard/oauth/refresh` API: the access token was refreshed
   by the backend without frontend involvement.

Also as in Canvas if you repeat the same test on _this_ branch the same
thing will happen but you will see a request to the refresh API in dev
tools.