Add useFetch hook to make data fetching in components easier
Whenever data is fetched for display in a component, there are a common
set of states that we need to keep track of and side effects that need
to be performed:

 - The render function needs to know whether a fetch is happening, or
   succeeded, or failed
 - If the inputs to the fetch change or the component is unmounted, the
   previous fetch needs to be canceled

Add a hook to encapsulate this. The API is intentionally very similar to SWR
[1]. That has a bunch of additional functionality that we don't yet need and
that users need to be aware of, but a future migration should be easy if needed.

[1] https://swr.vercel.app