Simplify grade fetching using `useFetch` hook
Use the `useFetch` hook to simplify fetching the grade for a student in
the SubmitGradeForm component. In particular `useFetch` handles ignoring
the results of fetches when the input query changes, avoiding the need
for the component to do it.

There was an error in a test where an async function was faked with a
stub that threw synchronously. This has been corrected to be async like
the real implementation.