Refactor the LTI 1.3 func tests to not rely on JS config on the error page
Currently we assume the error page has JS config, but it doesn't need any,
so the tests being this way prevent us from changing that.

We still want to use the error page as an indicator of whether you are a
student or a teacher, but this can be accomplished by some text matching.