Separate checking the GUID and updating AI data in the view
This is mostly "for show" to allow us to see what's happening. This allows a
 clean division between things we need for the view to work (checking the
GUID) and data storage for our persisted model.

The code here is the same / a little worse, as it calls the check guid twice
 effectively. But it makes the intent clearer.

This also avoids a potential problem where something crucial is happening as a
side effect of updating the data, but it's not clear. So you could easily
think you are doing something safe by moving the update, where you are actually
creating a huge problem.