Move launch data persistence to the constructor
This ensures it's performed on every launch. This actually was true
but each launch was responsible for making sure we called it.

This means the launch methods can concentrate on what they need to do
and not have to ensure this is called. It also makes it easier to
relocate.