Make JSTOR article thumbnail show first page
With the default `offset` parameter value of 0, the thumbnail API was
returning a thumbnail of the last page of the article instead of the
first. Setting the parameter to 1 returns the first page. Also increase
the width to provide a sharper image on higher-resolution displays. The
size of the thumbnail in the UI is such that it can be difficult to
discern if the expected article is being shown, but the layout needs
some changes to improve this.