Fix handling of JSTOR metadata/thumbnail requests that include a DOI prefix
Metadata/thumbnail requests for JSTOR article IDs with a DOI prefix
(eg. 10.5325/jafrideve.18.2.0019) did not match the expected route. It
seems this is because the path is URL-decoded before being matched
against the route's path pattern [1], and the default regex for path
parameters is `[^/]+`. Fix this by specifying a custom pattern for the
path parameter.

[1] Per https://docs.pylonsproject.org/projects/pyramid/en/latest/narr/urldispatch.html#route-pattern-syntax:

    > Literal strings in the path segment should represent the decoded value of the PATH_INFO provided to Pyramid