Show error in JSTOR article picker if item is a collection
PDFs are not available for collections, only specific articles in a
collection. For this reason, disable the Submit button in the JSTOR
article picker if the item is a collection and present an error to
explain why and how to resolve the issue.

The `is_collection` property of the article metadata API response is
just a placeholder for the moment, while some schema changes are worked
out on JSTOR's side.