Simplify the JSTOR service by observing that most methods do the same thing
In many methods there were the same chain of calls to various functions.
These have been gathered into one function that does all of them instead
of repeating the code in every method.

This also makes greater use of inlining in general for methods which
are only used in one place.