Automate finding requirements file prerequisites
Makefile currently contains a list of the prerequisite requirements
files for each requirements file that has prerequisites. If we add a new
requirements file with prerequisites or change the prerequisites of an
existing requirements file we may forget to update this list in
Makefile with confusing results.

Remove the hard-coded list of prerequisites from Makefile in favour of
getting make to find the prerequisites itself by finding the `-r` lines
in the requirements/*.in files.