Sync the user before we issue credentials in the Gateway
This means it's not possible for us to know that a user is "missing" in any way
as we will create it from the settings given. This also means we can't use
mock values as they will get updated.

As long as the caller issues the same LTI params as they got from the LMS this should
work fine.

Without this it's possible to issue credentials for a user which doesn't
exist yet.